requirements.txt
Django
Pillow
django-cors-headers
djangorestframework
